+++ R/reexports.R
@@ -25,3 +25,7 @@ hardhat::extract_preprocessor
 #' @importFrom hardhat extract_parameter_set_dials
 #' @export
 hardhat::extract_parameter_set_dials
+#'
+#' @importFrom hardhat extract_parameter_dials
+#' @export
+hardhat::extract_parameter_dials